+++ CHANGELOG.md
@@ -173,6 +173,7 @@
 - Nexo parser: added handling for "Transfer to Nexo User" and "Earn Interest Adjustment" in Details.
 - KuCoin parser: added "Fiat Deposits", "Third-Party Payment", "Fast Trade Orders" files.
 - Binance parser: added "Funding" Account, and "Send" and "Campaign Related Reward" Operations.
+- Coinbase parser: added "Donation" and "Admin Debit" transaction types.
 ### Changed
 - Conversion tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))
 - Accounting tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))